% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/qualified_teams.R
\docType{data}
\name{qualified_teams}
\alias{qualified_teams}
\title{Qualified Teams}
\format{
A data frame with 625 rows and 8 variables. See the package codebook
  (\code{variables}) for full column definitions.

A data frame with 8 variables:
\describe{
\item{key_id}{\code{integer}. The unique ID number for the observation.}
\item{tournament_id}{\code{text}. The unique ID number for the tournament.
References \code{tournament_id} in the \code{tournaments} dataset.}
\item{tournament_name}{\code{text}. The name of the tournament.}
\item{team_id}{\code{text}. The unique ID number for the team. References
\code{team_id} in the \code{teams} dataset.}
\item{team_name}{\code{text}. The name of the team.}
\item{team_code}{\code{text}. The 3-letter code for the team.}
\item{count_matches}{\code{integer}. The number of matches that the team
played in the tournament.}
\item{performance}{\code{text}. The furthest stage of the tournament
reached by the team.}
}
}
\source{
Original database by Joshua C. Fjelstul, Ph.D.
  \url{https://github.com/jfjelstul/worldcup}. Extended and maintained
  by Sadanand Venkataraman
  \url{https://github.com/sadanandv/fifa-worldcup-data}.
}
\usage{
qualified_teams

qualified_teams
}
\description{
This dataset records all qualified teams. There is one observation per team
per tournament. It includes the tournament, the team, and the performance of
each team (the furthest stage reached).

This dataset records all qualified teams. There is one observation per team
per tournament. It includes the tournament, the team, and the performance
of each team (the furthest stage reached). There are 8 variables and 625
observations.
}
\keyword{datasets}
